Replace bad chars in flag names
Radare2 changed the flag API as part of version 5.0.0. With this API,
flags with invalid characters are not created with a different name.
Instead, an error is thrown. This change adds a filter that removes
invalid characters in the flag name before the flag is created.

For more info about the change, see:
https://github.com/radareorg/radare2/pull/18020